Models
Added a bean model, cup, and void stuff

Created tutorial scaffold. You have to learn how to turn to get out of
this area.

Added some portal spinning things. Added them to where the beans spawn and
the void at the bottom of the world.
Particles spawn out of the void at the bottom of the world

Added some lighting

I have added and changed a bunch of stuff, but I don't know if I can say
that I've made it any better.