Fuel
Dozer now uses fuel as it moves. It can be refilled by buying a fuel tank.
Which extends from the DLC. Push it over to buy it
Now the game is going to be about pushing as many beans as possible off
the edge before you run out of fuel. Fuel jugs will be dropped off the
conveyor in the center over time and when you push off enough beans

Belts have a scrolling texture.

Added a portable fuel jug model