Submission Build
So.... It doesn't run in browser. Gets too intense.
Not sure if there's something I can do to fix that. Hard to get around the
fact that there are a ton of physics objects in the scene.
The little bean popup text things seemed very expensive so dialed those
back a lot.

Quality
I had to bump the quality up so the headlights would work.