Performance and Bean Counter UI
Made an object pool for the bean text objects. When you want to spawn a
new one, you just grab the oldest one instead and reuse it.

Made a counter for how many beans you've pushed and how many you need to
push for the next goal in the UI

TextMeshPro
Apparently the auto-adjusting font size is very expensive. I turned that
off because I don't need it for the text objects that always just say
'bean' and never change.
Technically, this increases performance, but the live html version just
has them completely disabled, so not really.

From 1 test, this build seems to run a bit better than the live version